+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 * `coco_time_combine` can now ignore seasons / handle median images with less memory
 * `use_grid_negatives` as dataset option, which can be set to "cleared" to only use negatives from cleared regions.
+* Add `modality_dropout` to kwcoco dataloader
 
 ### Changed
 
@@ -19,6 +20,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Started transition to a new package name: `geowatch`.
 * CI updated to `pyenv:3.11`
 * Cleanup `util_yaml`.
+* Lots of CLI improvements.
+* Faster loading of multiple kwcoco files in smartwatch stats, ensemble, combine.
 
 ### Fixed